requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
werkzeug==2.3.7
openai>=1.0.0
reportlab==4.0.4
python-dotenv==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.7
gunicorn==21.2.0 